chore(core): Consolidate pathbased and filesystem importers.
